refactored test_booking and debug
-I have made some logical errors involving detail page, payment page and submit, cancel needs some rework in order to accommodate for RentalRequest
submit --todo
renter can re-submit their rental request again after their previous rental request had been rejected, but it is currently overwriting over existing RentalRequest. This problem is to be addressed later.
if renter cannot cancel their rental during waiting for admin approval
