+++ renthub/tests/test_booking/test_payment_status.py
@@ -1,0 +1,5 @@
+"""Tests of booking: payment status."""
+
+from django.test import TestCase
+from django.urls import reverse
+from renthub.models import Rental, Room, Renter
