+++ renthub/views/submit_payment.py
@@ -41,10 +41,13 @@ def submit_payment(request, room_number):
         messages.info(request, "You already rented this room.")
         return render(request, "renthub/payment.html", {"room": room, "rental_exists": rental_exists})
 
-    # Create RentalRequest if it doesn't already exist
-    if not RentalRequest.objects.filter(room=room, renter=renter).exists():
-        RentalRequest.objects.create(room=room, renter=renter, price=room.price)
-        messages.success(request, f"Rental request for room {room_number} submitted successfully.")
+    latest_request = RentalRequest.objects.filter(renter=renter, room=room).order_by('-id').first()
+    if latest_request:
+        if latest_request.status != 'reject':
+            messages.warning(request, "You cannot submit a new rental request until the previous one is rejected.")
+            return render(request, "renthub/payment.html", {"room": room, "rental_exists": rental_exists})
+    rental_request = RentalRequest.objects.create(room=room, renter=renter, price=room.price)
+    messages.success(request, f"Rental request for room {room_number} submitted successfully.")
 
     # Proceed with rental logic
     rental_fee = room.price * 1  # Adjust if needed
